Batfish: support 4-byte as numbers
Mostly, int->long and Integer -> Long.

Also, stop warning about 4-byte OS in NX-OS extractor.

Also, update private AS range.

+++ projects/batfish/src/main/java/org/batfish/dataplane/ibdp/VirtualRouter.java
@@ -1250,8 +1250,8 @@ public class VirtualRouter extends ComparableStructure<String> {
         continue;
       }
 
-      int localAs = neighbor.getLocalAs();
-      int remoteAs = neighbor.getRemoteAs();
+      long localAs = neighbor.getLocalAs();
+      long remoteAs = neighbor.getRemoteAs();
       String remoteHostname = remoteIp.toString();
       String remoteVrfName = _vrf.getName();
       RoutingPolicy exportPolicy = _c.getRoutingPolicies().get(neighbor.getExportPolicy());
@@ -1376,7 +1376,7 @@ public class VirtualRouter extends ComparableStructure<String> {
           }
         }
         if (ebgpSession) {
-          SortedSet<Integer> newAsPathElement = new TreeSet<>();
+          SortedSet<Long> newAsPathElement = new TreeSet<>();
           newAsPathElement.add(localAs);
           transformedOutgoingRouteBuilder.getAsPath().add(0, newAsPathElement);
         }
